moved getCourse api call to api.ts, more search bar stuff
